Changed Linux makefile to use wx3.0.0 (and not treat warnings as errors - RenderWindow needs to be updated to avoid the use of depreciated wx methods); conversions between wxString and std::string modified across all files to be wx-unicode compliant; compiles and links OK, but currently crashing under GTK

+++ DataPlotter/src/utilities/math/matrix.cpp
@@ -1377,7 +1377,7 @@ double Matrix::ReduceToBidiagonalForm(Matrix &U, Matrix &V, Matrix &W, double *r
 
 		if (i < U.rows && i != V.rows - 1)
 		{
-			for (k = l - 1; k < (int)V.rows; k++)
+			for (k = l - 1; k < V.rows; k++)
 				scale += fabs(U.elements[i][k]);
 
 			if (scale != 0.0)
